# storage/api-docs/api-docs.yaml
...
openapi: 3.0.0
info:
  title: Development Challenge for PHP
  description: This is a sample API documentation for the Development Challenge for
    PHP
  version: 1.0.0
paths:
  "/api/actors":
    get:
      tags:
      - Actors
      summary: Display a listing of the actors
      operationId: 91a2ff1cae626a1a599ae2e37c48053c
      responses:
        '200':
          description: A listing of the actors
          content:
            application/json:
              schema:
                properties:
                  actors:
                    "$ref": "#/components/schemas/Actor"
                type: object
    post:
      tags:
      - Actors
      summary: Store a newly created actor in storage
      operationId: 49a96663e0b82b80d414aadfe3ce9646
      requestBody:
        description: Pass actor data
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                genre:
                  type: string
                birth_date:
                  type: string
              type: object
      responses:
        '200':
          description: Actor created successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  actor:
                    "$ref": "#/components/schemas/Actor"
                type: object
        '422':
          description: Invalid data passed
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  errors:
                    type: object
                type: object
...
  "/api/movies":
    get:
      tags:
      - Movies
      summary: Display a listing of the movie
      operationId: a7859dc5cd106bbeae5492e22df893ec
      responses:
        '200':
          description: A listing of the movie
          content:
            application/json:
              schema:
                properties:
                  movies:
                    "$ref": "#/components/schemas/Movie"
                type: object
    post:
      tags:
      - Movies
      summary: Store a newly created movie in storage
      operationId: a303e8fe5382b02a193fa994ebef739f
      requestBody:
        description: Pass movie data
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                release_date:
                  type: string
                synopsis:
                  type: string
                rating:
                  type: integer
                length:
                  type: string
                genre:
                  type: string
                director_id:
                  type: integer
                age_classification_id:
                  type: integer
                actors:
                  type: array
                  items:
                    type: integer
              type: object
      responses:
        '201':
          description: A newly created movie
          content:
            application/json:
              schema:
                properties:
                  movie:
                    "$ref": "#/components/schemas/Movie"
                type: object
        '404':
          description: Director not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
  "/api/movies/{movie}":
    get:
      tags:
      - Movies
      summary: Display the specified movie
      operationId: 36a46ca60c0936c9a2e4abd20099c5e5
      parameters:
      - name: movie
        in: path
        description: Movie id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: The specified movie
          content:
            application/json:
              schema:
                properties:
                  movie:
                    "$ref": "#/components/schemas/Movie"
                type: object
        '404':
          description: Not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
    put:
      tags:
      - Movies
      summary: Update the specified movie in storage
      operationId: a7e02fe8561bf4f720ddeea18767bdd8
      parameters:
      - name: movie
        in: path
        description: Movie id
        required: true
        schema:
          type: integer
      requestBody:
        description: Pass movie data
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                release_date:
                  type: string
                synopsis:
                  type: string
                rating:
                  type: integer
                length:
                  type: string
                genre:
                  type: string
                director_id:
                  type: integer
                age_classification_id:
                  type: integer
                actors:
                  type: array
                  items:
                    type: integer
              type: object
      responses:
        '200':
          description: The specified movie
          content:
            application/json:
              schema:
                properties:
                  movie:
                    "$ref": "#/components/schemas/Movie"
                type: object
        '404':
          description: Director not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
    delete:
      tags:
      - Movies
      summary: Remove the specified movie from storage
      operationId: 639d91f77a9b037a9083ee71509102da
      parameters:
      - name: movie
        in: path
        description: Movie id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Removed specified movie
          content:
            application/json:
              schema:
                properties:
                  movie:
                    "$ref": "#/components/schemas/Movie"
                type: object
        '404':
          description: Not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
...
components:
  schemas:
    GenderEnum:
      description: 'Gender options: Male, Female'
      type: string
      enum:
      - M
      - F
    GenreEnum:
      description: 'Genre options: Action, Comedy, Drama, Fantasy, Horror, Mystery,
        Romance, Science Fiction, Thriller'
      type: string
      enum:
      - AC
      - CO
      - DR
      - FA
      - HO
      - MY
      - RO
      - SF
      - TH
    Actor:
      title: Actor
      description: Actor model
      properties:
        id:
          description: Actor id
          type: integer
          example: '1'
        name:
          description: Actor name
          type: string
          example: John Doe
        birth_date:
          description: Actor birth date
          type: string
          format: date
          example: '1990-01-01'
        gender:
          "$ref": "#/components/schemas/GenderEnum"
      type: object
...
    Movie:
      title: Movie
      description: Movie model
      properties:
        id:
          description: Movie id
          type: integer
          example: '1'
        title:
          description: Movie title
          type: string
        release_date:
          description: Movie release date
          type: string
          format: date
        synopsis:
          description: Movie synopsis
          type: string
        rating:
          description: Movie rating
          type: integer
          example: '1'
        length:
          description: Movie length
          type: time
        genre:
          "$ref": "#/components/schemas/GenreEnum"
        director_id:
          description: Movie director id
          type: integer
          example: '1'
        age_classification_id:
          description: Movie age classification id
          type: integer
          example: '1'
        created_at:
          description: Movie created at
          type: string
          format: date-time
        updated_at:
          description: Movie updated at
          type: string
          format: date-time
      type: object
...